# openapi.yaml
openapi: "3.1.0"
info:
  title: Vyline Public API
  description: |
    Vyline セルフホスト型 LINE クライアントの公開 REST API。
    Bearer トークン認証が必要です（`VYLINE_API_ADMIN_SECRET` でトークンを作成）。

    ## クイックスタート

    1. `VYLINE_API_ADMIN_SECRET=<secret>` で Vyline を起動
    2. 管理者トークンを作成:
       ```
       curl -X POST http://localhost:3001/v1/tokens \
         -H "Authorization: Bearer <secret>" \
         -H "Content-Type: application/json" \
         -d '{"name":"my-bot","accountIds":["main"]}'
       ```
    3. 取得した `vyl_...` トークンを Bearer として使用
  version: "1.0.0"
servers:
  - url: "{baseUrl}/v1"
    variables:
      baseUrl:
        default: "http://localhost:3001"
        description: Vyline サーバー URL
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "vyl_<base64url>"
    adminSecret:
      type: http
      scheme: bearer
      bearerFormat: "<VYLINE_API_ADMIN_SECRET>"
  schemas:
    Error:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
...
paths:
  /accounts:
    get:
      summary: ログイン中アカウント一覧
      operationId: listAccounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: アカウントリスト
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        accountId:
                          type: string
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...